requirements.txt
fastapi
uvicorn
python-dotenv
requests
ollama
strands-agents
# Dependências de Teste
pytest
httpx
# Frontend
streamlit